исправлена проблема с истекшим токеном

+++ src/app/common/directives/favorite-item.directive.ts
@@ -8,8 +8,13 @@ import {
   signal,
 } from '@angular/core';
 import { Store } from '@ngrx/store';
-import { selectFavoriteById } from '../../store/user/user.selectors';
+import {
+  getUserStatus,
+  selectFavoriteById,
+} from '../../store/user/user.selectors';
 import { addFavorite, removeFavorite } from '../../store/user/user.actions';
+import { AuthStatus } from '../types/types';
+import { Router } from '@angular/router';
 
 @Directive({
   selector: '[appFavoriteItem]',
@@ -17,21 +22,30 @@ import { addFavorite, removeFavorite } from '../../store/user/user.actions';
 })
 export class FavoriteItemDirective implements OnInit {
   private store = inject(Store);
+  private router = inject(Router);
   public offerId = input.required<string>();
+  public isAuthorized = false;
 
   public isFavorite = signal(false);
 
   public ngOnInit(): void {
+    this.store.select(getUserStatus).subscribe((status) => {
+      this.isAuthorized = status === AuthStatus.Auth;
+    });
     this.store.select(selectFavoriteById(this.offerId())).subscribe((offer) => {
       this.isFavorite.set(offer?.isFavorite || false);
     });
   }
 
   @HostListener('click') onClick() {
-    if (this.isFavorite()) {
-      this.store.dispatch(removeFavorite({ favoriteId: this.offerId() }));
+    if (this.isAuthorized) {
+      if (this.isFavorite()) {
+        this.store.dispatch(removeFavorite({ favoriteId: this.offerId() }));
+      } else {
+        this.store.dispatch(addFavorite({ favoriteId: this.offerId() }));
+      }
     } else {
-      this.store.dispatch(addFavorite({ favoriteId: this.offerId() }));
+      this.router.navigate(['/login']);
     }
   }
